perf(vite): report the emitted module's shape from the native compiler
`generateOutput` needed six facts about the module the compiler had just
produced — whether it has a default export, whether it defines `_sfc_main`,
whether it exports a named `render`/`ssrRender`, and where the `export default`
keyword starts and ends — and recovered them by running `parseProgram` over that
same output. The #3425 profile measured ~77 ms of oxc work against ~49 ms for
the native compile batch it post-processes, and most of it is not the parse but
`jsonParseAst`, the cost of materializing oxc's AST as JavaScript objects across
the NAPI boundary.

The compiler now reports the shape, so that boundary crossing disappears: the
same parse runs natively and eight scalars cross instead of a whole AST.

The issue's plan was to record the offsets at each `export default` write site.
This derives them from the finished module instead, which is sound by
construction rather than by maintenance: two later passes rewrite the module
wholesale — `ensure_javascript_output` can reprint the entire program through
`Codegen`, and `finalize_output_mode` rewrites `export default` to `return` line
by line for `mode: "function"` — so every recorded offset would have to be
invalidated by both, and a stale offset is worse than an absent one because the
consumer splices at a byte position. Deriving after every pass leaves nothing to
invalidate.

It parses rather than scans for the same reason `analyzeFastDefaultOutput` bails
to a parse: the user's own `<script>` is concatenated into this output verbatim,
so `export default` can appear inside a string or a comment. Two tests pin that.

The JS side reads `compiled.moduleShape ?? analyzeModuleOutput(output)`. The
fallback is permanent, not transitional — a `.vpc` cache entry written before
this field existed reads back without it, and the rspack and unplugin builders
never set it — so no `PRECOMPILE_CACHE_FORMAT` bump is needed and #3399 can still
take the 2 -> 3 bump on its own, as agreed on both issues. A test pins both arms
against each other so they cannot drift.

Source-length budget: `lib.rs` was 553 lines, over the limit, so the module could
not be declared without splitting something. Its ~465 lines of inline tests move
to `parse_tests`, `compile_tests` and `template_tests`, split on the seam the
test names already draw; two insta snapshots are renamed to match their new
module path. `types.ts` was 401, so the batch types move to `batch-types.ts`.

Compiler output is unchanged: 6064 SFCs — vue-element-admin, vue2-elm, the repo's
fixture projects and the materialized ecosystem worktrees including misskey —
compile byte-identically against the pre-change baseline.

Refs #3425
