+++ crates/vize_atelier_sfc/src/lib.rs
@@ -43,6 +43,7 @@ pub mod compile_script;
 pub mod compile_template;
 pub mod croquis;
 pub mod css;
+pub mod module_shape;
 pub mod parse;
 pub mod rewrite_default;
 pub mod script;
@@ -83,471 +84,10 @@ pub use vize_atelier_core::CompilerError;
 pub use vize_atelier_dom::compile_template;
 
 #[cfg(test)]
-mod snapshot_tests;
-
+mod compile_tests;
 #[cfg(test)]
-mod tests {
-    use super::{SfcCompileOptions, compile_sfc, compile_sfc_with_template_syntax, parse_sfc};
-    use vize_atelier_core::TemplateSyntaxMode;
-    use vize_carton::config::VueVersion;
-
-    #[test]
-    fn template_compile_options_default_dialect_is_vue3() {
-        // The default per-file dialect must be modern Vue 3 — the zero-cost path.
-        let options = SfcCompileOptions::default();
-        assert_eq!(options.template.dialect, VueVersion::V3);
-    }
-
-    #[test]
-    fn dialect_threads_through_compile_without_changing_vue3_output() {
-        // PR2 is plumbing only: a non-Vue-3 dialect must reach the compile
-        // options (and therefore ParserOptions/TransformOptions), but no
-        // dialect-specific behavior is wired yet, so a plain Vue 3 template
-        // compiles byte-identically regardless of the selected dialect.
-        let source = r#"
-<template>
-  <div :class="cls" @click="onClick">{{ msg }}</div>
-</template>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-
-        let mut v2_options = SfcCompileOptions::default();
-        v2_options.template.dialect = VueVersion::V2;
-        assert_eq!(v2_options.template.dialect, VueVersion::V2);
-
-        let v3 = compile_sfc(&descriptor, SfcCompileOptions::default()).unwrap();
-        let v2 = compile_sfc(&descriptor, v2_options).unwrap();
-
-        assert_eq!(
-            v3.code, v2.code,
-            "dialect plumbing must not change Vue 3 codegen output"
-        );
-    }
-
-    #[test]
-    fn test_parse_simple_sfc() {
-        let source = r#"
-<template>
-  <div>Hello World</div>
-</template>
-
-<script>
-export default {
-  name: 'HelloWorld'
-}
-</script>
-
-<style>
-.hello { color: red; }
-</style>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-
-        assert!(descriptor.template.is_some());
-        assert!(descriptor.script.is_some());
-        assert_eq!(descriptor.styles.len(), 1);
-    }
-
-    #[test]
-    fn test_parse_script_setup() {
-        let source = r#"
-<template>
-  <div>{{ msg }}</div>
-</template>
-
-<script setup>
-import { ref } from 'vue'
-const msg = ref('Hello')
-</script>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-
-        assert!(descriptor.template.is_some());
-        assert!(descriptor.script_setup.is_some());
-    }
-
-    #[test]
-    fn test_parse_scoped_style() {
-        let source = r#"
-<template>
-  <div class="container">Scoped</div>
-</template>
-
-<style scoped>
-.container { background: blue; }
-</style>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-
-        assert_eq!(descriptor.styles.len(), 1);
-        assert!(descriptor.styles[0].scoped);
-    }
-
-    #[test]
-    fn test_compile_sfc_with_define_emits() {
-        let source = r#"
-<template>
-  <button @click="onClick">{{ count }}</button>
-</template>
-
-<script setup>
-import { ref } from 'vue'
-const emit = defineEmits(['update'])
-const count = ref(0)
-function onClick() {
-    emit('update', count.value)
-}
-</script>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-        let result = compile_sfc(&descriptor, SfcCompileOptions::default()).unwrap();
-        // `@vue/compiler-sfc` keeps the defineEmits runtime argument verbatim
-        // (single quotes from source), it does not re-serialize to double quotes.
-        assert!(
-            result.code.contains(r#"emits: ['update']"#),
-            "unexpected code:\n{}",
-            result.code
-        );
-        assert!(
-            result.code.contains("const emit = __emit"),
-            "unexpected code:\n{}",
-            result.code
-        );
-        assert!(
-            result.code.contains("emit('update', count.value)")
-                || result.code.contains("emit(\"update\", count.value)"),
-            "unexpected code:\n{}",
-            result.code
-        );
-
-        insta::assert_snapshot!(result.code.as_str());
-    }
-
-    #[test]
-    fn test_compile_sfc_standard_warns_for_invalid_html_self_closing() {
-        let source = r#"
-<template>
-  <div />
-  <span></span>
-</template>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-        let result = compile_sfc(&descriptor, SfcCompileOptions::default()).unwrap();
-
-        assert!(
-            result
-                .warnings
-                .iter()
-                .any(|warning| warning.message.contains("Invalid self-closing syntax")),
-            "expected invalid self-closing warning: {:?}",
-            result.warnings
-        );
-        assert!(result.code.contains("_createElementVNode(\"div\""));
-    }
-
-    #[test]
-    fn test_compile_sfc_strict_errors_for_invalid_html_self_closing() {
-        let source = r#"
-<template>
-  <div />
-</template>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-        let error = compile_sfc_with_template_syntax(
-            &descriptor,
-            SfcCompileOptions::default(),
-            TemplateSyntaxMode::Strict,
-        )
-        .expect_err("strict syntax should reject invalid self-closing HTML");
-
-        assert!(error.message.contains("Invalid self-closing syntax"));
-    }
-
-    #[test]
-    fn test_compile_sfc_define_model_with_type_args_preserves_body() {
-        // Regression test: defineModel<Type>('name', { opts }); was wrongly detected
-        // as a multi-line macro call because the line ends with `;` not `)`.
-        // This caused all subsequent setup code to be swallowed by the macro tracker.
-        let source = r#"
-<template>
-  <div>{{ fx }}</div>
-</template>
-
-<script setup lang="ts">
-interface Layer { fxId: string }
-const layer = defineModel<Layer>('layer', { required: true });
-
-const fx = layer.value.fxId;
-if (fx == null) {
-  throw new Error('not found');
-}
-</script>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-        let result = compile_sfc(&descriptor, SfcCompileOptions::default()).unwrap();
-
-        insta::assert_snapshot!(result.code.as_str());
-    }
-
-    #[test]
-    fn test_compile_sfc_imported_component_numeric_v_for() {
-        let source = r#"
-<template>
-  <Child
-    v-for="(id, index) in 4"
-    :key="id"
-    :label="String(index)"
-  />
-</template>
-
-<script setup lang="ts">
-import { Child } from "./components";
-</script>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-        let result = compile_sfc(&descriptor, SfcCompileOptions::default()).unwrap();
-
-        assert!(
-            result
-                .code
-                .contains(r#"import { Child } from "./components";"#),
-            "script setup import should be preserved. Got:\n{}",
-            result.code
-        );
-        assert!(
-            result.code.contains("_createBlock(_unref(Child)"),
-            "numeric component v-for should render the imported component. Got:\n{}",
-            result.code
-        );
-        assert!(
-            !result.code.contains(r#"_createElementVNode("Child""#),
-            "numeric component v-for must not render Child as a native element. Got:\n{}",
-            result.code
-        );
-    }
-
-    #[test]
-    fn test_compile_sfc_v_for_dynamic_prop_without_children_emits_null_children() {
-        let source = r#"
-<template>
-  <div
-    v-for="_, i in ary"
-    :prop="val"
-  ></div>
-</template>
-
-<script setup lang="ts">
-import { ref } from "vue"
-const ary = ref([])
-const val = ref(0)
-</script>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-        let result = compile_sfc(&descriptor, SfcCompileOptions::default()).unwrap();
-        let normalized = result.code.replace('\n', " ");
-
-        assert!(
-            normalized.contains(r#"_createElementBlock("div", { prop: val.value }, null, 8"#),
-            "v-for element with dynamic props and no children must emit a null children argument. Got:\n{}",
-            result.code
-        );
-        assert!(
-            !normalized.contains(r#"_createElementBlock("div", { prop: val.value }, 8"#),
-            "patch flag must not occupy the children argument. Got:\n{}",
-            result.code
-        );
-    }
-
-    #[test]
-    fn test_compile_sfc_ts_ref_condition_and_handler_keep_value_access() {
-        use vize_carton::ToCompactString;
-
-        let source = r#"
-<template>
-  <div>
-    <template v-if="folder == null">
-      <MkButton @click="isRootSelected = true" />
-    </template>
-    <template v-else>
-      <MkButton
-        v-if="!selectedFolders.some(f => f.id === folder!.id)"
-        @click="selectedFolders.push(folder)"
-      />
-      <MkButton
-        v-else
-        @click="selectedFolders = selectedFolders.filter(f => f.id !== folder!.id)"
-      />
-    </template>
-  </div>
-</template>
-
-<script setup lang="ts">
-import { ref } from 'vue'
-
-const folder = ref<{ id: string } | null>(null)
-const selectedFolders = ref<{ id: string }[]>([])
-const isRootSelected = ref(false)
-</script>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-        let script_setup = descriptor.script_setup.as_ref().unwrap();
-        let template = descriptor.template.as_ref().unwrap();
-        let croquis = crate::script::analyze_script_setup_to_summary(&script_setup.content);
-        let mut binding_metadata = crate::BindingMetadata::default();
-        binding_metadata.is_script_setup = croquis.bindings.is_script_setup;
-        for (name, binding_type) in croquis.bindings.iter() {
-            binding_metadata
-                .bindings
-                .insert(name.to_compact_string(), binding_type);
-        }
-        for (local, key) in &croquis.bindings.props_aliases {
-            binding_metadata
-                .props_aliases
-                .insert(local.to_compact_string(), key.to_compact_string());
-        }
-
-        let template_output = crate::compile_template::compile_template_block(
-            template,
-            &crate::TemplateCompileOptions::default(),
-            crate::compile_template::TemplateBlockCompileContext {
-                scope_id: "",
-                apply_scope_id: false,
-                has_scoped: false,
-                is_ts: true,
-                inline: true,
-                component_name: None,
-                bindings: Some(&binding_metadata),
-                croquis: Some(croquis),
-            },
-            vize_atelier_core::TemplateSyntaxMode::Standard,
-            &vize_atelier_core::CodegenOptions::default(),
-        )
-        .expect("template compile should succeed");
-        let template_code = template_output.code;
-        assert!(
-            template_code.contains("!selectedFolders.value.some((f) => f.id === folder.value.id)"),
-            "unexpected template code:\n{}",
-            template_code
-        );
-        assert!(
-            template_code.contains("$event => (selectedFolders.value = selectedFolders.value.filter((f) => f.id !== folder.value.id))"),
-            "unexpected template code:\n{}",
-            template_code
-        );
-        let (_imports, _hoisted, _preamble, render_body, _render_fn_name) =
-            crate::compile_template::extract_template_parts(&template_code);
-        assert!(
-            render_body.contains("!selectedFolders.value.some((f) => f.id === folder.value.id)"),
-            "unexpected render body:\n{}",
-            render_body
-        );
-        assert!(
-            render_body.contains("$event => (selectedFolders.value = selectedFolders.value.filter((f) => f.id !== folder.value.id))"),
-            "unexpected render body:\n{}",
-            render_body
-        );
-
-        let result = compile_sfc(&descriptor, SfcCompileOptions::default()).unwrap();
-
-        assert!(
-            result
-                .code
-                .contains("!selectedFolders.value.some((f) => f.id === folder.value.id)"),
-            "unexpected code:\n{}",
-            result.code
-        );
-        assert!(
-            result
-                .code
-                .contains("selectedFolders.value = selectedFolders.value.filter((f) => f.id !== folder.value.id)"),
-            "unexpected code:\n{}",
-            result.code
-        );
-        assert!(!result.code.contains("($event) => (($event) =>"));
-    }
-
-    #[test]
-    fn test_compile_sfc_nested_custom_directive_keeps_inline_with_directives() {
-        let source = r#"
-<template>
-  <div>
-    <button
-      v-show="ok"
-      v-appear="shouldEnableInfiniteScroll ? fetchOlder : null"
-      @click="fetchOlder"
-    >
-      Load more
-    </button>
-  </div>
-</template>
-
-<script setup lang="ts">
-const ok = true
-const shouldEnableInfiniteScroll = true
-const fetchOlder = () => {}
-</script>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-        let result = compile_sfc(&descriptor, SfcCompileOptions::default()).unwrap();
-        let normalized = result.code.replace('\n', " ");
-
-        assert!(
-            result
-                .code
-                .contains(r#"const _directive_appear = _resolveDirective("appear")"#),
-            "unexpected code:\n{}",
-            result.code
-        );
-        assert!(
-            normalized.contains(
-                r#"[_directive_appear, shouldEnableInfiniteScroll ? fetchOlder : null], [_vShow, ok]"#
-            ),
-            "unexpected code:\n{}",
-            result.code
-        );
-        assert!(
-            result.code.contains("_withDirectives(_createElementVNode(")
-                && result.code.contains("\"button\""),
-            "unexpected code:\n{}",
-            result.code
-        );
-    }
-
-    #[test]
-    fn test_compile_sfc_native_v_model_keeps_custom_directive() {
-        let source = r#"
-<template>
-  <input v-model="local" v-example @input="touches++">
-</template>
-
-<script setup lang="ts">
-import { ref } from 'vue'
-
-const local = ref('initial')
-const touches = ref(0)
-</script>
-"#;
-        let descriptor = parse_sfc(source, Default::default()).unwrap();
-        let result = compile_sfc(&descriptor, SfcCompileOptions::default()).unwrap();
-        let normalized = result.code.replace('\n', " ");
-
-        assert!(
-            result
-                .code
-                .contains(r#"const _directive_example = _resolveDirective("example")"#),
-            "unexpected code:\n{}",
-            result.code
-        );
-        assert!(
-            normalized.contains(r#"[_vModelText, local.value], [_directive_example]"#),
-            "unexpected code:\n{}",
-            result.code
-        );
-        assert!(
-            result.code.contains(r#""onUpdate:modelValue":"#)
-                && result.code.contains("local.value = $event"),
-            "unexpected code:\n{}",
-            result.code
-        );
-    }
-}
+mod parse_tests;
+#[cfg(test)]
+mod snapshot_tests;
+#[cfg(test)]
+mod template_tests;
